Added nullValueList (nvl) function
The nvl function takes two parameters.  Its purpose is to evaluate
the first parameter and return it when it's not null or undefined.
If the first parameter is null or undefined it returns the first
non-null value in the second parameter.  The second parameter can
be a single value or an array of values.

Unit tests were added to test simple cases of the function.  The
package version was updated to reflect the new functionality.