adding new functionality for dlogger.vals
the new functionality allows the function to handle two new kinds of
inputs.  both new inputs simplify the process of displaying variable
information.

one of the new input formats allows the use of an array of arrays.
each sub-array must have 2 values.  the first value is the key and
the second value is the corresponding data which we want to show.

example: [['keyName','data'],['otherKey', 123]]

the second new input format allows the user of an object with a single
property which acts as the key and the value.  this is similary to the
key/value pair signature, but more succint.

example: [{val: 'data'}, {second: 'property'}]