adding on/off functions and method chaining functionality
two new functions where added: on/off.  as their name implies the
functions turn output on or off by setting the _debug variable to
true or false respectively.

the new method chaining functionality was implemented to allow
easier execution of specific output.  for example, if we want to
output only one specific line or lines we could do something like:

logger.on().lob('message').off();

this can be better for performance and readability when we need to
troubleshoot one specific function or value and we don't want the
entire project to output data.